over

+++ src/main/java/com/jiao/chatonlineserver/services/servicesImpl/UserServicesImpl.java
@@ -1,7 +1,9 @@
 package com.jiao.chatonlineserver.services.servicesImpl;
 
+import com.jiao.chatonlineserver.pojo.PostMessage;
 import com.jiao.chatonlineserver.pojo.User;
 import com.jiao.chatonlineserver.pojo.UserTest;
+import com.jiao.chatonlineserver.repository.PostMessageMapper;
 import com.jiao.chatonlineserver.repository.UserMapper;
 import com.jiao.chatonlineserver.services.UserService;
 import org.apache.ibatis.annotations.Param;
@@ -15,10 +17,14 @@ import java.util.List;
 public class UserServicesImpl implements UserService {
     @Resource
     UserMapper userDao;
+    @Resource
+    PostMessageMapper pmm;
+
     @Override
     @Transactional
     public void insertUser(User user) {
         if (user!=null){
+            user.setRoleID("user");
             this.userDao.insertUser(user);
         }else{
             System.out.println("user=="+user);
@@ -38,8 +44,11 @@ public class UserServicesImpl implements UserService {
     }
 
     @Override
-    public void deleteUser(@Param("user") User user) {
-        if (user!=null){
+    @Transactional
+    public void deleteUser(@Param("user") User user, PostMessage pm) {
+        if (user!=null && pm!=null){
+            //先删除从表中的数据在删除主表的
+            this.pmm.delPostByUid(pm);
             this.userDao.deleteUser(user);
         }
